pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "myrag"
description = "use of rag for personal use"
version = "1.0"
dependencies = [
    "pip-tools",
    "python-dotenv",
    "pydantic",
    "langchain",
    "pymongo[srv]==3.12"
]

[tools.setuptools.packages.find]
where = ["src"]